refactor: modularize large files and fix all clippy warnings (#66)
Completed comprehensive refactoring to address GitHub issue #66 by breaking down
all files exceeding 700 lines into focused, cohesive modules following the
project's architectural principles.

## Refactored modules (6 files → 41 modules):

- `crates/cache/src/security/audit.rs` (768 lines → 8 modules)
  Split into: config, context, events, integrity, logger, types, mod

- `crates/utils/src/resilience/circuit.rs` (749 lines → 8 modules)
  Split into: config, metrics, retry, state, transitions, types, tests, mod

- `crates/task/src/builder.rs` (745 lines → 6 modules)
  Split into: validation, env_expansion, dependency, security, conversion, mod

- `crates/cli/src/secrets.rs` (715 lines → 4 modules)
  Split into: resolver, manager, tests, mod

- `crates/core/src/types.rs` (711 lines → 8 modules)
  Split into: environment, commands, capabilities, security, files, tasks, shared, mod

- `crates/cache/src/streaming.rs` (711 lines → 7 modules)
  Split into: reader, writer, operations, path_utils, finalization, tests, mod

## Fixed clippy warnings:

- `result_unit_err`: Added proper error types instead of Result<_, ()>
- `ptr_arg`: Changed &PathBuf to &Path in function signatures
- `too_many_arguments`: Created TaskExecutionParams struct for functions with 11+ params
- `module_inception`: Renamed test modules to avoid name conflicts with parent modules
- `type_complexity`: Created type alias for complex Arc<RwLock<HashMap>> types
- `uninlined_format_args`: Fixed format string interpolation
- `redundant_closure`: Simplified closure expressions

All tests pass (468 tests), zero clippy warnings, and nix flake check succeeds.
