+++ crates/core/src/types/shared.rs
@@ -1,0 +1,6 @@
+//! Shared types and utilities used across different domains
+
+use std::sync::Arc;
+
+/// Shared string type for immutable strings
+pub type SharedString = Arc<str>;
